.hpp fix

+++ shramko.maksim/S1/functionsForList.hpp
@@ -13,7 +13,7 @@ namespace shramko
   using PairList = ForwardList< std::pair< std::string, NumberList >>;
 
   void printNames(const PairList& lists, std::ostream& out);
-  bool checkSumOverflow(size_t a, size_t b);
+  bool checkSumOverflow(int a, int b);
   void printSumResult(const ForwardList< int >& sums, bool overflow, std::ostream& out);
   void processLists(const PairList& lists, size_t maxLen, bool& overflow, std::ostream& out);
 }
